refactor: move everything into source directory
