+++ src/ws_server/ws_server.go
@@ -1,7 +1,7 @@
 package ws_server
 
 import (
-	"akv/akv"
+	"akv/src/akv"
 	"io"
 	"log"
 	"net/http"
