+++ src/ws_server/ws_server_interface.go
@@ -1,7 +1,7 @@
 package ws_server
 
 import (
-	"akv/akv"
+	"akv/src/akv"
 	"github.com/gorilla/websocket"
 )
